review: split gate_travel.rs to gate_travel/{mod,tests}.rs + bump sentinel off 0x7000_0700 collision
Addresses Copilot's two new concerns:

1. The combined file grew to 665 lines, over the 500-line soft cap.
   Converted to a directory module: gate_travel/mod.rs holds the
   production handle_gate_travel (219 lines) and gate_travel/tests.rs
   holds the test fixtures and 5 tests (450 lines). Same split pattern
   as damage_apply, livewire, mail.

2. The live-DB test's sentinel TEST_ACCOUNT/CHAR_A/CHAR_B used
   0x7000_0700/0701/0702 — colliding with vendor/paid_repair which
   already uses 0x7000_0700. Concurrent runs (or one test's cleanup
   spilling into the other's fixture state) could surface as cross-
   test interference. Bumped to 0x7000_0D00 — a free slot. Documented
   the slot map inline so future additions don't repeat the collision.
